requirements.txt
# Core Flask Dependencies
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Image Processing
Pillow==10.0.1
opencv-python-headless==4.8.1.78

# Machine Learning - Core
numpy==1.24.3
scipy==1.11.3

# ML Frameworks (uncomment what you need)
scikit-learn==1.3.0
# tensorflow==2.13.0
# torch==2.0.1
# torchvision==0.15.2
# xgboost==1.7.6
# lightgbm==4.1.0

# Data Processing
pandas==2.0.3
joblib==1.3.2

# System Monitoring
psutil==5.9.6

# Optional: ONNX Support
# onnxruntime==1.16.0
# onnxruntime-gpu==1.16.0  # For GPU support

# Optional: Advanced Image Processing
# albumentations==1.3.1
# imgaug==0.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0